pyproject.toml
[project]
name = "motif"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai>=1.57.0",
    "ollama>=0.6.1",
    "openai>=2.15.0",
    "pandas>=2.3.3",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-levenshtein>=0.27.3",
    "requests>=2.32.5",
    "thefuzz>=0.22.1",
    "tqdm>=4.67.1",
    "uvicorn>=0.40.0",
]
